add working defer test with label
